#338 Update to attachment form vue files
